requirements.txt
python-telegram-bot==13.15
requests
pandas
numpy
ta
python-dateutil